[JSC] Fix a !binding.isScratch() assertion failure on 32-bit https://bugs.webkit.org/show_bug.cgi?id=298157
Reviewed by NOBODY (OOPS!).

In 32-bit, we use scratch fp regs to check the bounds of the input of in the
truncSaturated and truncTrapping functions, before emitting the right ccall.

However, the scratch fp registers are still bound when we emit the ccall, which
leads to this assertion failure. These assertions only fail in debug mode and
still provide the correct result in release mode because these fp registers are
not used in the emitted ccall.

This PR changes the ScratchScope in the truncSaturated and truncTrapping
functions, so it's destroyed and the fp registers are unbound before we emit
the ccalls.

* Source/JavaScriptCore/wasm/WasmBBQJIT32_64.cpp:
(JSC::Wasm::BBQJITImpl::BBQJIT::truncTrapping):
(JSC::Wasm::BBQJITImpl::BBQJIT::truncSaturated):